+++ packages/demo/scripts/stage.ts
@@ -3,7 +3,7 @@ import { appsConfig } from "../src/config/apps.config";
 import { ANSI } from "../../core/src/constants";
 import { NAME, TAG_LINE } from "../src/config/branding.config";
 import { BASE_URL, BUILD_DIR, DOMAIN } from "../src/config/deploy.config";
-import { skin } from "../src/config/skin.config";
+import { defaultSkin } from "../src/config/skin.config";
 import { name } from "../package.json";
 
 const PATHS = {
@@ -17,7 +17,7 @@ function generateSitemapXml() {
 	const date = new Date();
 	const lastModified = `${date.getFullYear()}-${date.getMonth() + 1}-${date.getDate()}`;
 
-	const images = skin.wallpapers.map((path) => `
+	const images = defaultSkin.wallpapers.map((path) => `
 		<image:image>
 			<image:loc>${BASE_URL.slice(0, -1) + path}</image:loc>
 		</image:image>`