+++ packages/demo/src/Main.tsx
@@ -1,11 +1,33 @@
-import { ReactElement } from "react";
+import { ReactElement, useEffect, useState } from "react";
 import { Router } from "./router/Router";
-import { skin } from "./config/skin.config";
+import { defaultSkin } from "./config/skin.config";
 import { NAME, TAG_LINE } from "./config/branding.config";
 import { appsConfig } from "./config/apps.config";
-import { ProzillaOS } from "prozilla-os";
+import { getViewportParams, ProzillaOS } from "prozilla-os";
+import { macOsSkin, minimalSkin, pixelSkin } from "@prozilla-os/skins";
 
 export function Main(): ReactElement {
+	const [skin, setSkin] = useState(defaultSkin);
+
+	useEffect(() => {
+		const params = getViewportParams();
+
+		if (params.skin == null)
+			return;
+
+		switch (params.skin) {
+			case "mac":
+				setSkin(macOsSkin);
+				break;
+			case "minimal":
+				setSkin(minimalSkin);
+				break;
+			case "pixel":
+				setSkin(pixelSkin);
+				break;
+		}
+	}, []);
+
 	return <ProzillaOS
 		systemName={NAME}
 		tagLine={TAG_LINE}