+++ packages/apps/terminal/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prozilla-os/terminal
 
+## 1.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @prozilla-os/core@1.2.1
+
 ## 1.1.1
 
 ### Patch Changes