+++ packages/apps/text-editor/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prozilla-os/text-editor
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @prozilla-os/core@1.2.1
+
 ## 1.1.0
 
 ### Minor Changes